requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
motor==3.6.0
pandas==2.2.3
python-dotenv==1.0.1
